third_party/blink/web_tests/platform/generic/fast/dom/Range/compareBoundaryPoints-2-expected.txt

Test compareBoundaryPoints results in a particular case, see bug 20738.

Some text.


START_TO_START = 1
START_TO_END = 1
END_TO_START = -1
END_TO_END = -1

PASS
